Updated vpn connecting script to check for the type of vpn client needed

+++ vitwifi/vit-vpn-connect.sh
@@ -1,3 +1,19 @@
 nm-applet &
+
+# setting default value for the client parameter
+${client:-"proton"}
+
+if [[ $client == "proton" ]]; then
 pkill protonvpn
 protonvpn-cli connect --fastest
+
+elif [[ $client == 'windscribe' ]]; then
+pkill windscribe
+windscribe-cli connect best
+windscribe-cli disconnect
+windscribe-cli connect best
+
+else
+  echo "Chose --client windscribe or proton"
+
+fi
